+++ CHANGELOG.md
@@ -109,6 +109,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `party` event - new optional `amount` of maximal affected players
 - `drop` event
 - `itemdurability` event, condition and variable
+- `numbercompare` condition
 - Things that are also added in 1.12.X:
     - new line support for `journal_lore` in `messages.yml`
     - FastAsyncWorldEdit compatibility